requirements.txt
fastapi[all]>=0.100.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.25.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
redis>=5.0.0
aiomysql>=0.2.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
celery>=5.3.0
